requirements.txt
paho-mqtt
face-recognition
pillow